refactor: 显式化隐式类型转换，消除 816 条编译警告
把编译器提示的隐式窄化转换改为显式 static_cast，并修掉几处会静默截断的转换与
日志格式串不匹配，使 CI 编译日志只剩非自有第三方驱动的存量警告。

- pxview-render：527 处隐式窄化（Qt 的 qreal 坐标传入 QPainter 整型接口、采样点
  计数在 uint64_t 与 double 之间换算等）改为显式 static_cast，取值与绘制结果逐位
  不变。该层此前已开启 -Wconversion 严格检查、构建注释里也声明基线为 0，实际却长期
  残留这 527 条，新引入的转换问题会被完全淹没；现已真正归零，该检查恢复为可用的
  防回归闸门。
- pxview-core/pxview-data：修复 4 处新回归 —— API 服务把 std::stoul 的结果交给
  g_variant_new_uint32 时未做范围转换（64 位平台上静默丢弃高 32 位）；文件名提示
  解析把 QString::lastIndexOf 返回的 qsizetype 存进 int，超长路径会被截断。
- pxlogic/demo 驱动：日志格式串与实参类型不匹配（%llu 打 uint64_t/gint64、%zu 打
  int32_t）在 64 位平台上属未定义行为，改用 PRIu64 / G_GINT64_FORMAT / %d 精确对应；
  另清理 demo 驱动中的死变量与未使用参数。
- libsigrok：C++ 专用选项 -fvisibility-inlines-hidden 不再传给 C 编译单元（该库为
  纯 C），消除 Linux 上每个翻译单元一条、累计 273 条的 cc1 警告。

未改动的部分：其余第三方硬件驱动（sipeed-slogic-analyzer、raspberrypi-pico、
siglent-sds、dreamsourcelab-dslogic、asix-sigma、chronovu-la、ftdi-la、
ikalogic-scanaplus、pipistrello-ols、devantech-eth008、scpi-*、serial-lcr、
rdtech-dps、ipdbg-la）以及 common/minizip、libsigrokdecode 一律保持原样 —— 未经硬件
实测不改动硬件相关代码，它们约 106 条警告属已知存量。

验证：全量 ninja -j 16 通过、0 error；cd build && ctest 为 38/39，唯一失败的
test_v_offset_animation（7 个 MakeWay 用例）系 05d29dcc 重构 make_way.h 时遗留的既有
失败，已用 git stash 暂存本提交全部改动后复跑确认，与本提交无关。
更新说明.txt / doc/NEWS25 / doc/NEWS31 三份已同步（字节级一致）。

+++ PXView/pv/api/session_service.cpp
@@ -5041,7 +5041,7 @@ if (!root_decoder || !root_decoder->decoder())
                         else if (g_variant_is_of_type(opt_def->def, G_VARIANT_TYPE("i")))
                             val = g_variant_new_int32(std::stoi(opt.second));
                         else if (g_variant_is_of_type(opt_def->def, G_VARIANT_TYPE("u")))
-                            val = g_variant_new_uint32(std::stoul(opt.second));
+                            val = g_variant_new_uint32(static_cast<guint32>(std::stoul(opt.second)));
                         else if (g_variant_is_of_type(opt_def->def, G_VARIANT_TYPE("t")))
                             val = g_variant_new_uint64(std::stoull(opt.second));
                         else
@@ -5067,7 +5067,7 @@ if (!root_decoder || !root_decoder->decoder())
                     else if (g_variant_is_of_type(opt_def->def, G_VARIANT_TYPE("i")))
                         val = g_variant_new_int32(std::stoi(opt.second));
                     else if (g_variant_is_of_type(opt_def->def, G_VARIANT_TYPE("u")))
-                        val = g_variant_new_uint32(std::stoul(opt.second));
+                        val = g_variant_new_uint32(static_cast<unsigned int>(std::stoul(opt.second)));
                     else if (g_variant_is_of_type(opt_def->def, G_VARIANT_TYPE("t")))
                         val = g_variant_new_uint64(std::stoull(opt.second));
                     else
